% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amerika_colors.R
\docType{data}
\name{amerika_palettes}
\alias{amerika_palettes}
\title{American politics-inspired color palettes}
\format{An object of class \code{list} of length 5.}
\usage{
amerika_palettes
}
\description{
Use \code{\link{amerika_palette}} to produce desired color palette
}
\keyword{datasets}
